1. Spring SecurityConfig 설정 변경
2. 객체 기본키 타입을 int -> long 타입으로 변경

- int 타입에 비해 long 타입이 호환성, 범용성, 안정성 등이 높아지기 때문에 변경

+++ src/main/java/com/cafeHi/www/board/controller/QnaPageController.java
@@ -54,7 +54,7 @@ public class QnaPageController {
 		CustomUser userInfo = (CustomUser) principal;
 	    Member getMember = userInfo.getMember();
 	  
-	    int member_code = getMember.getMember_code();
+	    Long member_code = getMember.getMember_code();
 
 	    Map<String , Object> myPageQnAMap = new ConcurrentHashMap<>();
