+++ src/main/java/com/cafeHi/www/order/service/OrderMenuService.java
@@ -8,7 +8,7 @@ public interface OrderMenuService {
 
     void insertOrderMenu(OrderMenu orderMenu);
 
-    List<OrderMenu> findOrderMenuList(int member_code);
+    List<OrderMenu> findOrderMenuList(Long member_code);
 
     void cancelOrderMenu(OrderMenu orderMenu);
 }
