+++ src/main/java/com/cafeHi/www/order/service/OrderService.java
@@ -3,7 +3,7 @@ package com.cafeHi.www.order.service;
 import com.cafeHi.www.order.dto.Orders;
 
 public interface OrderService {
-    int insertOrder(Orders order);
+    Long insertOrder(Orders order);
 
     void cancelOrder(Orders order);
